requirements.txt
flask==2.0.1
python-dotenv==0.19.0
google-generativeai==0.3.0
gunicorn==20.1.0
fastf1==2.3.0
pandas==1.5.3
numpy==1.24.3
plotly==5.14.1
scipy==1.10.1